Upgraded to GL-based Static API
Renamed SnapshotOptions to ClassicSnapshotOptions. Added a similar SnapshotOptions class tailored to the GL-based Static API. Split out ClassicSnapshotTests.
